requirements.txt
fastapi
openai
pydantic
python-dotenv
uvicorn
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
openai>=1.50.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
typer>=0.12.0,<1.0.0
